refactor: replace Next.js custom server with standalone Socket.IO process
Custom server approach was incompatible with Vercel and disabled Turbopack.
Separate the Socket.IO server into an independent process (socket-server/)
to align with the architecture doc (Railway/Render deployment).

- Add socket-server/index.ts: standalone Node.js + Socket.IO server on port 3001
  - Exposes POST /internal/emit for Next.js API routes to broadcast events
  - Handles room:join / room:leave / chat:send with Prisma persistence
  - Auth middleware via NextAuth session cookie verification
  - Optional SOCKET_INTERNAL_SECRET for internal endpoint security
- Add lib/socket-emitter.ts: HTTP client to call socket-server from API routes
- Update leave/close routes to use emitToRoom() instead of global io instance
- Restore next dev --turbopack; use concurrently for combined dev startup
- Remove server.ts and lib/socket.ts (custom server approach)
