+++ package.json
@@ -3,7 +3,9 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "NODE_ENV=development tsx server.ts",
+    "dev": "concurrently --names \"next,socket\" --prefix-colors \"blue,green\" \"next dev --turbopack --port 3000\" \"tsx socket-server/index.ts\"",
+    "dev:next": "next dev --turbopack --port 3000",
+    "dev:socket": "tsx socket-server/index.ts",
     "build": "next build",
     "start": "next start",
     "lint": "eslint",
@@ -20,6 +22,7 @@
     "@prisma/client": "^7.4.1",
     "@tanstack/react-query": "^5.90.21",
     "clsx": "^2.1.1",
+    "dotenv": "^17.3.1",
     "lucide-react": "^0.469.0",
     "next": "16.1.6",
     "next-auth": "5.0.0-beta.30",
@@ -35,6 +38,7 @@
     "@types/pg": "^8.16.0",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "concurrently": "^9.2.1",
     "eslint": "^9",
     "eslint-config-next": "16.1.6",
     "prisma": "^7.4.1",
